Refactor player death code to use Unity Events
Although the Killable script probably won't be used for anything other than the player at this point, player character specific enums and responsibilities had seeped in.

This killable script pretty much just acts as an interface and fires off and event if wired up.

This makes it easier to chain other actions on in the future such as enabling UI elements on death.